Rearrange Storybook, add subheaders for better orientation
